requirements.txt
blinker==1.9.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
et_xmlfile==2.0.0
filelock==3.20.0
Flask==3.1.2
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
license-expression==30.4.4
markdown-it-py==4.0.0
markup==0.2.2
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
numpy==2.3.5
openpyxl==3.1.5
packageurl-python==0.17.5
packaging==25.0
pandas==2.3.3
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.5.0
py-serializable==2.1.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rich==14.2.0
six==1.17.0
sortedcontainers==2.4.0
toml==0.10.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xlsxwriter==3.2.9
